Rename to shortcut.css
Hack.css didn't really illustrate the underlying principles of this library and limited it to the context of hackathons. However, my goals were simply to create something that was quickly understandable. I wanted a user who may not have any knowledge of CSS, glance at a class name, and understand how to use it and what it will do -- just like a shortcut. Thus, it only made sense to name it shortcut.css.